+++ src/oms_sensemaking/cocom/__init__.py
@@ -1,0 +1,1 @@
+"""COCOM Traversal Sensemaker"""